[VM] Remove InvokeJit overhead in VM (#347)
* sync

* update

* finish tvm refactor

* update cudnn & cublas op dispatch

* add openv cache in vm

* fix rebase err

* update

* update

* clean up

* tweak

* add more tests for vm

* update test

* lint

* fix

* lint

* lint

* tweak

* lint

* lint

* fix cpptest

* fix

* x

* fix test case

+++ tests/python/op/ty/test_type_reduce.py
@@ -54,7 +54,7 @@ def test_reduce(op, shape, keepdims, dtype):
     else:
         ty_shape = t_y.shape
     if not same:
-        y_ty = TensorType(ty_shape, dtype="int64")
+        y_ty = TensorType(ty_shape, dtype="int32")
     else:
         y_ty = TensorType(ty_shape, dtype=dtype)
     desired_type = FuncType([x_ty], y_ty)
@@ -101,7 +101,7 @@ def test_reduce_all_any(op, shape, keepdims, dtype):
     t_y = torch_fwd(t_x, dim=axis, keepdim=keepdims)
     x_ty = TensorType(t_x.shape, dtype=dtype)
     if not same:
-        y_ty = TensorType(t_y.shape, dtype="int64")
+        y_ty = TensorType(t_y.shape, dtype="int32")
     else:
         y_ty = TensorType(t_y.shape, dtype=dtype)
     desired_type = FuncType([x_ty], y_ty)
